Refactor backend to DDD structure
