+++ backend/src/domain/entities/product.ts
