+++ backend/src/infrastructure/repositories/product.repository.ts
@@ -1,5 +1,5 @@
 import { createClient } from "../database/dbClient";
-import { Product } from "../types/product";
+import { Product } from "../../domain/entities/product";
 
 // get all products
 const getAllProducts = async () => {
